feat: create staff profile and staff auth

+++ src/main/java/com/g42/platform/gms/staff/profile/infrastructure/repository/RoleJpaRepo.java
@@ -3,5 +3,5 @@ package com.g42.platform.gms.staff.profile.infrastructure.repository;
 import com.g42.platform.gms.staff.profile.infrastructure.entity.RoleJpa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleJpaRepo extends JpaRepository<RoleJpa,Long> {
+public interface RoleJpaRepo extends JpaRepository<RoleJpa,Integer> {
 }
